api_evaluation/requirements.txt
exa-py>=1.0.0
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
